nice! @curried + test=true works now
