+++ packages/doctor-general-jest/log.js
@@ -1,3 +1,9 @@
 import { complextrace } from 'envtrace'
 
-export const log = complextrace('doctor-general-jest', ['read', 'write'])
+export const log = complextrace('doctor-general-jest', [
+  'read',
+  'write',
+  'verbose',
+  'renderer',
+  'curried',
+])
