+++ packages/kiddo/autotests/kiddo.spec.js
@@ -1,10 +1,18 @@
 // This test automatically generated by doctor-general.
 // Sourced from 'kiddo.js', edits to this file may be erased.
-import { signal } from '../kiddo'
+import { signal, exec } from '../kiddo'
+
+import { fork } from 'fluture'
+test('exec', done => {
+  const blah = Math.round(Math.random() * 100000)
+  fork(done)(z => {
+    expect(z.stdout).toEqual('' + blah)
+    done()
+  })(exec('echo', [blah]))
+})
 
 import { pipe, map } from 'ramda'
 import { readFile } from 'file-system'
-import { fork } from 'fluture'
 test('signal', done => {
   const cancel = () => {}
   pipe(
